requirements.txt
langchain==0.0.274
python-dotenv==1.0.0
gradio==3.41.2
openai==0.27.9
regex==2023.8.8 
tiktoken==0.4.0
#chromadb==0.4.7
pypdf==3.15.4
docx2txt==0.8
faiss-cpu==1.7.4
pinecone-client==2.2.2
boto3==1.28.40